% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ct_data}
\alias{ct_data}
\title{Sample simulated clinical trial dataset}
\format{
A data frame with 120 rows and 6 variables:
\describe{
\item{trt_group}{A character type variable used to specify the treatment
conditions of the subject}
\item{sex}{A character type variable used to specify the sex of the
subject}
\item{age}{A numeric type variable for specifying the age of the subject}
\item{race}{A character type variable used to specify the race of the
subject}
\item{pre_treatment_bp}{A numeric type variable for specifying the
systolic blood pressure of the patient before study enrollment}
\item{post_treatment_bp}{A numeric type variable for specifying the
systolic blood pressure of the patient after study enrollment}
}
}
\source{
{Created through simulation to serve as an example}
}
\usage{
ct_data
}
\description{
A simulated dataset containing variables for testing out ALPFunctions
functions
}
\examples{
data(ct_data)
}
\keyword{datasets}
