% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_flex_table.R
\name{gt_flex_table}
\alias{gt_flex_table}
\title{Create publication style table using gtsummary and flextable}
\usage{
gt_flex_table(
  data_frame,
  group_by = NULL,
  label_list = list(),
  digits_list = list(),
  value_list = list(),
  type_list = list(),
  include = everything(),
  add_p = FALSE,
  add_p_test = NULL,
  add_p_test_args = NULL,
  add_p_group = NULL,
  add_total_col = FALSE,
  header_text = "**Table 1: Insert Table Title**",
  return_object = "flextable"
)
}
\arguments{
\item{data_frame}{A dataframe object}

\item{group_by}{A categorical variable to group the table by}

\item{label_list}{A list object to specify for variable labels}

\item{digits_list}{A list object to specify for digits lengths}

\item{value_list}{A list object to specify summary statistics and values}

\item{type_list}{A list object to specify variable types}

\item{add_p}{If TRUE, adds statistic p-values}

\item{add_p_test}{If not NULL, specify the test statistic type}

\item{add_p_test_args}{If not NULL, specify test statistic arguments}

\item{add_p_group}{If not NULL, specify the variable used for ID grouping for paired data}

\item{add_total_col}{If TRUE, adds the totals for all columns}

\item{header_text}{Specify header}

\item{return_object}{Specifies what object}
}
\value{
Returns a flextable or gtsummary table object
}
\description{
This function utilizes the gtsummary tbl_summary function to summarize data
with custom settings and inputs and can output a flextable object with custom
inputs for rendering Rmarkdown word or html reports
}
