% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gt_to_flex_convert.R
\name{gt_to_flex_convert}
\alias{gt_to_flex_convert}
\title{Convert a gtsummary table to a flextable with custom designs}
\usage{
gt_to_flex_convert(
  gt_tbl,
  font_name = "Albany AMT",
  font_size_header = 12,
  font_size_body = 10,
  font_size_footer = 10
)
}
\arguments{
\item{gt_tbl}{A gtsummary object}

\item{font_name}{A character string to specify the font type for the
flextable, defaults to "Albany AMT"}

\item{font_size_header}{A numeric value to specify the font size for the
table header for the flextable, defaults to 12}

\item{font_size_body}{A numeric value to specify the font size for the
table body for the flextable, defaults to 10}

\item{font_size_footer}{A numeric value to specify the font size for the
table footer for the flextable, defaults to 10}
}
\value{
A flextable object
}
\description{
Convert a gtsummary table to a flextable with custom designs
}
