% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paired_ct_data}
\alias{paired_ct_data}
\title{Sample simulated clinical trial dataset with paired data}
\format{
A data frame with 120 rows and 6 variables:
\describe{
\item{participant_id}{A numeric variable used to uniquely identify each
subject}
\item{treat_time}{A character type variable used to specify the treatment
time of the observation for a subject}
\item{sex}{A character type variable used to specify the sex of the
subject}
\item{age}{A numeric type variable for specifying the age of the subject}
\item{race}{A character type variable used to specify the race of the
subject}
\item{bp}{A numeric type variable for specifying the
systolic blood pressure of the patient}
}
}
\source{
{Created through simulation to serve as an example}
}
\usage{
paired_ct_data
}
\description{
A simulated paired dataset containing variables for testing out ALPFunctions
functions
}
\examples{
data(paired_ct_data)
}
\keyword{datasets}
